// src/app/global/configs/routes/api/account.ts
export const enum ACCOUNT_API {

  getById = 'account/fetchBy/id/',
  fetchAll = 'account/fetch/all',
  create = 'account/new',
  createList = 'account/new/list',
  update = 'account/update/credentials',
  updateCredentials = 'account/update/credentials',
  deleteById = 'account/deleteBy/id/',
  deleteList = 'account/delete/list'

}